requirements.txt
Flask==3.1.0
openai==0.10.2
requests==2.25.1
python_dotenv
